@@ -18,14 +18,13 @@ package cmd
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 
 	"github.com/bitnami-labs/kubewatch/config"
 	"github.com/bitnami-labs/kubewatch/pkg/client"
 	"github.com/bitnami-labs/kubewatch/pkg/event"
-	"github.com/sirupsen/logrus"
+	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 )
 
@@ -61,7 +60,7 @@ Tests handler configs present in ~/.kubewatch.yaml by sending test messages`,
 		fmt.Println("Testing Handler configs from .kubewatch.yaml")
 		conf, err := config.New()
 		if err != nil {
-			logrus.Fatal(err)
+			log.Fatal().Err(err)
 		}
 		eventHandler := client.ParseEventHandler(conf)
 		e := event.Event{
@@ -94,7 +93,7 @@ var configViewCmd = &cobra.Command{
 Display the contents of the contents of ~/.kubewatch.yaml`,
 	Run: func(cmd *cobra.Command, args []string) {
 		fmt.Fprintln(os.Stderr, "Contents of ~/.kubewatch.yaml")
-		configFile, err := ioutil.ReadFile(filepath.Join(os.Getenv("HOME"), kubewatchConfigFile))
+		configFile, err := os.ReadFile(filepath.Join(os.Getenv("HOME"), kubewatchConfigFile))
 		if err != nil {
 			fmt.Fprintf(os.Stderr, "%v\n", err)
 			os.Exit(1)
