@@ -18,7 +18,7 @@ package cmd
 
 import (
 	"github.com/bitnami-labs/kubewatch/config"
-	"github.com/sirupsen/logrus"
+	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 )
 
@@ -30,7 +30,7 @@ var hipchatConfigCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		conf, err := config.New()
 		if err != nil {
-			logrus.Fatal(err)
+			log.Fatal().Err(err)
 		}
 
 		token, err := cmd.Flags().GetString("token")
@@ -39,7 +39,7 @@ var hipchatConfigCmd = &cobra.Command{
 				conf.Handler.Hipchat.Token = token
 			}
 		} else {
-			logrus.Fatal(err)
+			log.Fatal().Err(err)
 		}
 		room, err := cmd.Flags().GetString("room")
 		if err == nil {
@@ -47,11 +47,11 @@ var hipchatConfigCmd = &cobra.Command{
 				conf.Handler.Hipchat.Room = room
 			}
 		} else {
-			logrus.Fatal(err)
+			log.Fatal().Err(err)
 		}
 
 		if err = conf.Write(); err != nil {
-			logrus.Fatal(err)
+			log.Fatal().Err(err)
 		}
 	},
 }
