@@ -18,7 +18,7 @@ package cmd
 
 import (
 	"github.com/bitnami-labs/kubewatch/config"
-	"github.com/sirupsen/logrus"
+	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 )
 
@@ -30,7 +30,7 @@ var slackwebhookConfigCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		conf, err := config.New()
 		if err != nil {
-			logrus.Fatal(err)
+			log.Fatal().Err(err)
 		}
 
 		channel, err := cmd.Flags().GetString("channel")
@@ -39,7 +39,7 @@ var slackwebhookConfigCmd = &cobra.Command{
 				conf.Handler.SlackWebhook.Channel = channel
 			}
 		} else {
-			logrus.Fatal(err)
+			log.Fatal().Err(err)
 		}
 
 		username, err := cmd.Flags().GetString("username")
@@ -48,7 +48,7 @@ var slackwebhookConfigCmd = &cobra.Command{
 				conf.Handler.SlackWebhook.Username = username
 			}
 		} else {
-			logrus.Fatal(err)
+			log.Fatal().Err(err)
 		}
 
 		emoji, err := cmd.Flags().GetString("emoji")
@@ -64,11 +64,11 @@ var slackwebhookConfigCmd = &cobra.Command{
 				conf.Handler.SlackWebhook.Slackwebhookurl = slackwebhookurl
 			}
 		} else {
-			logrus.Fatal(err)
+			log.Fatal().Err(err)
 		}
 
 		if err = conf.Write(); err != nil {
-			logrus.Fatal(err)
+			log.Fatal().Err(err)
 		}
 	},
 }
