@@ -17,7 +17,7 @@ limitations under the License.
 package cmd
 
 import (
-	"github.com/sirupsen/logrus"
+	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
 )
 
@@ -35,8 +35,8 @@ var versionCmd = &cobra.Command{
 }
 
 func versionPrettyString() {
-	logrus.Info("gitCommit: ", gitCommit)
-	logrus.Info("buildDate: ", buildDate)
+	log.Info().Msgf("gitCommit: %s", gitCommit)
+	log.Info().Msgf("buildDate: %s", buildDate)
 }
 
 func init() {
